Clue States
Represent the states a clue can be in as enums in the abstract Clue. This means we can move the isActive(), isSolved(), activate(), and solve() methods to Clue and override them from subclasses as needed. Currently some functionality is determined from the Type attribute, which is a pretty bad code smell. Added a TODO for fixing it.