GeofenceManager Serialization Fix
When the scavenger hunt is loaded from a file or deserialized we now give its clues a new GeofenceManager so we don't have to worry about the fact that it wasn't serialized. Also fixed the progress bar not displaying correctly- app:srcCompat seems to be broken, use android:src instead as a quick and dirty fix.