Compass Clues and Notification Class
Added compass clues. You need to give the activity a serializable intent of a CompassClue, but we haven't done that yet. Also moved sending notifications out of GeofenceTransitionsIntentService so it can be done by other classes. Also did a bunch of minor misc changes that I forgot to commit last night.