Force Clue Solving
You can force solve a clue by clicking it 50 times. This does not reload the page but it does send a notification which will reload the page if you click it. Also reloaded the main page when you launch a pre-loaded scavenger hunt from the menu. Also removed the compass touch stuff from CompassActivity since it is no longer relevant.