ScavengerHuntDatabase Integration
Not gonna lie, this should have been several commits. Replaced FileUtil's load and save methods with ones for ScavengerHuntDatabase, and then fixed all the build failures and runtime exceptions that I could find. This required a lot of functionality to be added to ScavengerHuntDatabase, and a good amount to be added to ScavengerHunt.