QR Scanning, Almost
Added an activity for scanning QR codes and some code around it. Error handling isn't near where it needs to be (I think uploads and downloads silently fail) and the saved database is weird with activities loading/unloading, but with a little work we are close to a merge with master.