Manually Check Internet Connection
Manually check for internet access before trying to access firebase. Had to disable offline mode in firebase to prevent instances where the user got both a successful and error popup. There is definitely a better way that keeps the offline mode in firebase but it's not really high priority since the app doesn't need internet access for the vast majority of the time. Solves issue #17.